// src/types/planets.tsx
export interface Planets {
  name: string;
  rotation_period: number;
  orbital_period: number;
  diameter: number;
  climate: string;
  gravity: number;
  terrain: string;
  surface_water: string;
  population: number;
  residents: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}